Techdebt: Move imports around
