+++ mazure/azure_services/login/responses.py
@@ -8,7 +8,8 @@ from urllib.parse import unquote
 
 import jwt
 
-from mazure.mazure_core import MazureRequest, ResponseType
+from mazure.mazure_core import ResponseType
+from mazure.mazure_core.mazure_request import MazureRequest
 from mazure.mazure_core.route_mapping import register
