+++ mazure/azure_services/management/resource_groups/responses.py
@@ -3,7 +3,8 @@ import random
 import re
 import string
 
-from mazure.mazure_core import MazureRequest, ResponseType
+from mazure.mazure_core import ResponseType
+from mazure.mazure_core.mazure_request import MazureRequest
 from mazure.mazure_core.route_mapping import register
 
 from . import model
