+++ mazure/azure_services/management/subscriptions/responses.py
@@ -1,7 +1,8 @@
 import json
 import re
 
-from mazure.mazure_core import MazureRequest, ResponseType
+from mazure.mazure_core import ResponseType
+from mazure.mazure_core.mazure_request import MazureRequest
 from mazure.mazure_core.route_mapping import register
 
 from . import sub_model
