+++ mazure/azure_services/storage/responses.py
@@ -1,7 +1,8 @@
 import re
 from datetime import datetime
 
-from mazure.mazure_core import MazureRequest, ResponseType
+from mazure.mazure_core import ResponseType
+from mazure.mazure_core.mazure_request import MazureRequest
 from mazure.mazure_core.route_mapping import register
 
 from .models.blob import Blob
