+++ mazure/mazure_core/service_discovery.py
@@ -2,7 +2,8 @@ import re
 from gzip import compress
 from typing import Any, Dict
 
-from mazure.mazure_core import MazureRequest, ResponseType
+from mazure.mazure_core import ResponseType
+from mazure.mazure_core.mazure_request import MazureRequest
 
 from .route_mapping import registered_services  # pylint: disable
