+++ tests/tests_python/test_models/test_graph_applications.py
@@ -1,5 +1,3 @@
-from unittest import TestCase
-
 from peewee import SqliteDatabase
 
 from mazure.azure_services.management.graph.models import graph_backend
@@ -11,23 +9,23 @@ from mazure.azure_services.management.graph.models.application import (
 MODELS = [Application, DeletedApplication]
 
 # use an in-memory SQLite for tests.
-test_db = SqliteDatabase(":memory:")
+_test_db = SqliteDatabase(":memory:")
 
 
-class TestGraphApplications(TestCase):
-    def setUp(self) -> None:
+class TestGraphApplications:
+    def setup_method(self) -> None:
         # Bind model classes to test db. Since we have a complete list of
         # all models, we do not need to recursively bind dependencies.
-        test_db.bind(MODELS, bind_refs=False, bind_backrefs=False)
+        _test_db.bind(MODELS, bind_refs=False, bind_backrefs=False)
 
-        test_db.connect()
-        test_db.create_tables(MODELS)
+        _test_db.connect()
+        _test_db.create_tables(MODELS)
 
-    def tearDown(self) -> None:
-        test_db.drop_tables(MODELS)
+    def teardown_method(self) -> None:
+        _test_db.drop_tables(MODELS)
 
         # Close connection to db.
-        test_db.close()
+        _test_db.close()
 
     def test_applications_lifecycle(self) -> None:
         assert not graph_backend.list_applications()
