+++ tests/tests_python/test_requests/storage/test_blob_storage.py
@@ -32,25 +32,25 @@ MODELS = [
 ]
 
 # use an in-memory SQLite for tests.
-test_db = SqliteDatabase(":memory:")
+_test_db = SqliteDatabase(":memory:")
 
 
 class TestStorageAccounts:
-    def setup(self) -> None:
+    def setup_method(self) -> None:
         # Bind model classes to test db. Since we have a complete list of
         # all models, we do not need to recursively bind dependencies.
-        test_db.bind(MODELS, bind_refs=False, bind_backrefs=False)
+        _test_db.bind(MODELS, bind_refs=False, bind_backrefs=False)
 
-        test_db.connect()
-        test_db.create_tables(MODELS)
+        _test_db.connect()
+        _test_db.create_tables(MODELS)
 
         # create group
 
-    def teardown(self) -> None:
-        test_db.drop_tables(MODELS)
+    def teardown_method(self) -> None:
+        _test_db.drop_tables(MODELS)
 
         # Close connection to db.
-        test_db.close()
+        _test_db.close()
 
     def test_create_container(self) -> None:
         # Create StorageAccount
